libmlx5: Fix MLX5_SCATTER_TO_CQE environment variable handling.
When MLX5_SCATTER_TO_CQE==1, scatter cqe mode should be enabled.
Current code just does opposite of it when env is not set.

This fix returns it correctly, that is,
when env is set, and if MLX5_SCATTER_TO_CQE is set to 1, it
indicates to use scatter cqe, otherwise not.

Signed-off-by: Parav Pandit <parav@mellanox.com>