oracle: Add buildrpm and link spec file
THIS SHOULD NOT BE SUBMITTED UPSTREAM. DO NOT SQUASH!

'buildrpm' is required by the Oracle auto build system. Rather than
deviate from the upstream convention of having a per-distro directory
for distro-specific files, we add buildrpm and symlink the RPM spec
file so the build system can find it witin the 'oracle' directory.

buildrpm/
└── ol7
    └── rdma-core.spec -> ../../oracle/rdma-core.spec

An enhancement request has been opened to allow specifying the location
of the spec file rather than defaulting to 'buildrpm'. This provides
better alignment with the methods used by upstream projects. If/when
that ER is implemented, this commit will be reverted.

The packages derived from this project will only be built for OL7 and
will be restricted to OL7 with the UEK5 kernel. The rdma-core service
(replacement for the older rdma service) does not include support for
SysV initscripts. See commit 7ebed9d65a53 for more information.

In the future, we may make some or all of the individual packages (e.g.,
libibverbs, librdmacm, etc.) that are generated from this project
compatible with both OL7+UEK5 and OL6 and OL7 with the UEK4 kernel.
However, UEK5-based systems would rely on the rdma-core service and
UEK4-based systems would rely on the rdma service.

THIS SHOULD NOT BE SUBMITTED UPSTREAM. DO NOT SQUASH!

Orabug: 26829960

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>