oracle/spec: Add Epoch to package versioning and adjust Obsoletes
Add Epoch to package versioning:

FIX THIS BEFORE PUTTING PACKAGES IN ULN!

The Epoch tag is added to the spec file to avoid build-time and
install-time conflicts between versions of packages intended for
a particular UEK release or conflicts with RedHat-provided packages.

For build-time, multiple "BuildRequires" tags are used to express
contstraints using the epoch when applicable. This is because the
current tools do not support "BuildConflicts" or the newer Boolean
operators for expressing rich dependencies.

At this time, there are no modifications to "BuildRequires" for the RDMA
Core packages.

For run-time, the tags: "Conflicts", "Obsoletes", "Provides", and
"Requires" are updated, with the epoch, when needed.

Adjust Obsoletes:

FIX THIS BEFORE PUTTING PACKAGES IN ULN!

These changes need further study and are likely incorrect and/or
incomplete!

For now, remove "Obsoletes" tags for those packages which we currently
build and ship as part of oracle-ofed-release. This needs to be revisted
in the context of our build system, which mixes Epoch 4 (OL7+UEK4) and
Epoch 5 (OL7+UEK5) in the same yum channel as well as in the context of
run-time when packages are installed and different yum channels may or
may not be enabled or contain or exclude certain packages.

FIX THIS BEFORE PUTTING PACKAGES IN ULN!

Orabug: 27774601

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>