oracle: Load RDS when starting rdma service
Remove redundant checks for presence of RDS modules such that modules
will load when uncompressed (UEK4) or compresssed (UEK5) without having
to check for the presence of <module>.ko or <module>.ko.xz. The
"load_modules()" function already takes care of not loading modules if
they are not present.

"load_modules() can be further optimized to remove the "modinfo" loop
and use "modeprobe -q". This script will be abandoned in favor for the
kernel-boot init mechansisms so "load_modules()" will be left as-is.

Orabug: 27003153

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>
Reviewed-by: Sudhakar Dindukurti <sudhakar.dindukurti@oracle.com>