oracle/init: Load Resilient RDMA IP module
Load the Resilient RDMA IP module (resilient_rdmaip) when rdma.service
is started.

This also loads IPoIB by default instead of a conditional load (e.g., if
RDMAIP is enabled or RDS is enabled then load IPoIB). This is because
all of the InfiniBand ULPs present at this time depend on the IP address
of IPoIB for their functionality.

Previously, conditional loading of IPoIB was done because RDS was using
an API defined by IPoIB. However, even without this API dependency,
IPoIB must be loaded before RDS as it uses the IP address of the IPoIB.

A better approach, using systemd, should be considered in the future:

  1. Load IPoIB iff IB hardware is present. This needs to be done by the
     generic rdma-core kernel boot mechanism.

  2. Once that is done, we can conditionally load Resilient RDMA IP
     because it has an API dependency on IPoIB. Loading IPoIB will be
     required even on a RoCE-only system when using this module.

Orabug: 27740657

Signed-off-by: Sudhakar Dindukurti <sudhakar.dindukurt@oracle.com>
Reviewed-by: Avinash Repaka <avinash.repaka@oracle.com>
Signed-off-by: Aron Silverton <aron.silverton@oracle.com>