libmlx5: Disable MLX5_QP_FLAG_SCATTER_CQE for Shared-PD
MLX5_INLINE_SCATTER_{32,64} is an optimization inside the MLX5 driver,
that is triggered by the presence of flag MLX5_QP_FLAG_SCATTER_CQE.

It's best to understand first how SHPD works to understand the
underlying issue:

With Oracle's SHPD-hack, a memory region registered by process#0 is
shared by other processes (i.e. mapped into their address space).

That mapping is not guaranteed to land on the same virtual address in
process#1... and even "encouraged" to land on a different virtual
addresses if sysctl kernel.randomize_va_space is > 0.

What the client process does it calculate the delta between the address
that processs#0 created the shared segment at, and the address that said
shared memory segment landed on in its own address space. It then adds
that delta to the virtual addresses used in verbs such as
"ibv_post_recv".

Let's say process #0 has a shared memory segment at virtual address
0x2000 and process #1 has the same shared memory segment at 0x1000.

So process #1 would now "fake" the address by adding "0x1000" to each
request. Assuming it is only the HCA that uses said addresses, and the
memory-translation-table inside the HCA would make it land on the
correct address, as it doesn't really care which process is submitting
this request (it only checks for the PD to be correct... hence
shared-pd).

That works for as long as ONLY the HCA is using the address. If process
case (such as in this bug report), or in the worst-case it'll overwrite
some memory it doesn't really own.

So we simply turn off the MLX5_INLINE_SCATTER_{32,64} feature for QPs
created on behalf of PDs that are considered "inherited", e.g.  as
a result of calling "ibv_share_pd".

Orabug: 27949058

Signed-off-by: Gerd Rausch <gerd.rausch@oracle.com>
Reviewed-by: Ajaykumar Hotchandani <ajaykumar.hotchandani@oracle.com>
Signed-off-by: Aron Silverton <aron.silverton@oracle.com>