oracle/spec: Fix ldconfig warnings when uninstalling
When removing the libibverbs, libibumad, and librdmacm packages their
RPM POSTUN scriptlet throws a warning that, while proven to be harmless,
is annoying.

An example of the warning:

/sbin/ldconfig: relative path `0' used to build cache
warning: %postun(libibverbs-5:17.1-1.0.2.el7.x86_64) scriptlet failed, exit status 1
Non-fatal POSTUN scriptlet failure in rpm package

The transaction does actually succeed and the libraries are removed as
expected.

The warning is caused by comments in the spec file that confuse RPM when
it interprets the scriptlets.

  rpm -qp --scripts libibverbs-15-7.el7_5.x86_64.rpm
  postinstall program: /sbin/ldconfig
  postuninstall program: /sbin/ldconfig

but it is really:

  rpm -qp --scripts libibverbs-17.1-1.0.2.el7.x86_64.rpm
  postinstall program: /sbin/ldconfig
  postuninstall scriptlet (using /sbin/ldconfig):

  # libibumad

where the "# libibumad" is being included as part of the scriptlet.

Orabug: 28146600

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>