oracle/spec: Use Ninja (ninja-build) for building
The RDMA Core project recommends that developers use Ninja. In keeping
with that recommendation we will also use Ninja when packaging.

Oracle Linux does not ship Ninja (ninja-build) but it is available in
the EPEL repository. The build system must be configured to enable EPEL
when building this package.

Orabug: 28305731

Signed-off-by: Aron Silverton <aron.silverton@oracle.com>